+++ lib/rubocop/cop/chef/deprecation/resource_overrides_provides_method.rb
@@ -18,9 +18,7 @@ module RuboCop
   module Cop
     module Chef
       module ChefDeprecations
-        # Some providers in resources override the provides? method, used to check whether they are a valid
-        # provider on the current platform. In Chef Infra Client 13, this will cause an error if the provider
-        # does not also register themselves with the provides call.
+        # Some providers in resources override the provides? method, used to check whether they are a valid provider on the current platform. In Chef Infra Client 13, this will cause an error. Instead use the 'provides' method to register the provider.
         #
         # @example
         #
@@ -31,10 +29,6 @@ module RuboCop
         #
         #   # good
         #   provides :SOME_PROVIDER_NAME
-        #
-        #   def provides?
-        #    true
-        #   end
         #
         class ResourceOverridesProvidesMethod < Cop
           MSG = "Don't override the provides? method in a resource provider without also calling provides :SOME_PROVIDER_NAME. This will cause failures in Chef Infra Client 13 and later.".freeze